+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog
 
 **Next release**
 
+* Fixed photos for allowed contacts ([issue](/../../issues/1588))
 * Updated Slovenian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)